Fix: partial revert of  change to link-ptbs-to-dblsqd GH Actions workflow (#8062)
#### Brief overview of PR changes/additions
Revert a change I made in #8015 where I changed a `bash` variable name
into a SCREAMING_SNAKE_CASE one - unfortunately I did not (and still
don't) fully understand how it was processed by the `jq` (command line
JSON query tool) and broke the processing of the DBLSQD feed data and
thus stopped the script from working as intended.

#### Motivation for adding to Mudlet
Get Windows PTBs produced again.

#### Other info (issues closed, discussion etc)
This PR should put the affected bit of the script back to exactly how it
was prior to the PR that broke things!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>